change weatheboost return to double

+++ HeroesDB/Entity/Hero.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; } 
         public string Name { get; set; } = null!;
         public Powers Powers { get; set; }
-        public int Weatherboost { get; set; }
-        public int PowerLevel { get; set; } 
+        public double Weatherboost { get; set; }
+        public double PowerLevel { get; set; } 
         public bool IsVillain { get; set; }
     }
 }
